% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreIt.R
\name{exploreIt}
\alias{exploreIt}
\title{Generate Exploration Report}
\usage{
exploreIt(input_data, output_file = "exploreIt.html",
  output_format = "html", output_dir = getwd(), ...)
}
\arguments{
\item{input_data}{Data}

\item{output_file}{Name for output file}

\item{output_format}{Format for file, PDF or HTML}

\item{output_dir}{Directory to save the output file}
}
\description{
This function generates a report to explore the data
}
