% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepIt.R
\name{p.continuous}
\alias{p.continuous}
\title{Numeric Continuous Variables}
\usage{
p.continuous(dta, max_levels = 0.1)
}
\arguments{
\item{dta}{Data}

\item{max_levels}{Ratio of distinct values}
}
\description{
Numeric Continuous Variables
}
