% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepIt.R
\name{p.replaceNull}
\alias{p.replaceNull}
\title{Replace Null Values}
\usage{
p.replaceNull(dta, columns, null_value = "Missing")
}
\arguments{
\item{dta}{Data}

\item{columns}{Fields}

\item{null_value}{Value to populate for null fields}
}
\description{
Function generates plots to describe null values
}
