% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepIt.R
\name{p.structure}
\alias{p.structure}
\title{Detailed Data Structure}
\usage{
p.structure(dta, max_levels = 0.1)
}
\arguments{
\item{dta}{Data}

\item{max_levels}{Limits factors that are compared.}
}
\description{
Shows a more custom view of the data
}
