% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepIt.R
\name{p.summary}
\alias{p.summary}
\title{Describe Data}
\usage{
p.summary(dta)
}
\arguments{
\item{dta}{Data being loaded}
}
\value{
A data frame
}
\description{
This function adds to the describe function in the psych pacakage
}
